fix: relocate repository interfaces to Application layer - Moved repositories to Application/Interfaces to adhere to Clean Architecture principles - Removed Domain/Repositories folder
